sysddefs
When using fte cc to compile the progs.src i had to add a sysdefs.qc. It would fail without it so i essentially copied what was in the progs.sh file and made it a sysdefs.qc . This allowed me to just recompile the progs.src in FTE cc.